Fixes from review: Move platform stuff to fs_darwin...
and platformLookup method signature match Lookup.
